requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
astor=0.8.1=pypi_0
backcall=0.1.0=pypi_0
ca-certificates=2020.1.1=0
cachetools=4.1.0=pypi_0
certifi=2020.4.5.1=py37_0
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
facenet-pytorch=2.2.9=pypi_0
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
google-auth=1.14.3=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.29.0=pypi_0
h5py=2.10.0=pypi_0
idna=2.9=pypi_0
importlib-metadata=1.6.0=pypi_0
imutils=0.5.3=pypi_0
ipython=7.14.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.0=pypi_0
joblib=0.15.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.2.2=pypi_0
matplotlib=3.2.1=pypi_0
ncurses=6.2=he6710b0_1
numpy=1.18.4=pypi_0
oauthlib=3.1.0=pypi_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=h7b6447c_0
opt-einsum=3.2.1=pypi_0
parso=0.7.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.2=pypi_0
pip=20.0.2=py37_3
prompt-toolkit=3.0.5=pypi_0
protobuf=3.12.0=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=pypi_0
readline=8.0=h7b6447c_0
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=46.2.0=py37_0
six=1.14.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.31.1=h62c20be_1
tensorboard=2.0.2=pypi_0
tensorflow-estimator=2.0.1=pypi_0
tensorflow-gpu=2.0.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.0.0=pypi_0
tk=8.6.8=hbc83047_0
torch=1.5.0=pypi_0
torchvision=0.6.0=pypi_0
traitlets=4.3.3=pypi_0
urllib3=1.25.9=pypi_0
wcwidth=0.1.9=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
